+++ app/(after-login)/(simple)/(back)/post/[id]/page.tsx
