+++ app/(befor-login)/(simple)/(back)/(in-preparation)/layout.tsx
